Channel Scaling Refactor and addition to meter wizard
